Move the logic to a separate file
